Get rid of poster/mysimulate, and move the respective bits back to poster/arwm and poster/eval.
The function poster/mysimulate existed because it integrated poster/arwm and poster/eval with both calling a nested function xxlogpost. The function xxlogpost became a private method, mylogpost, and hence there's no benefits of integrating the back-ends for arwm and eval within one file.

Create a new function poster/mylogpoststruct that prepares a struct for use in poster/mylogpost; this new function is called from withing poster/arwm and poster/eval.

Change the name of the input argument `Draw` in the backend code for poster/arwm to `NDraw`. Change the name of that variable also in `poster/stats` when `saveEvery=` is used.